Move processing logic to own file
